Remove specifying calib repo in the command lines
If we use the 04_butler manual step to create the combined repo,
calibration repo is not in the butler configuration.  The only
way I found to correct that is to manually edit repositoryCfg.yaml
CmdLineTask checks this butler cfg; RuntimeError unless matching.

The alternative it to remove "--calib /datasets/hsc/calib/20200115"
from the command lines.